VCV-26: clean up code and some UI
